Moving abif_test.py into new testscripts directory
Also fixing up some of the tests and comments.  I know how to use
pytest SLIGHTLY better than when I wrote my original test suite,
and so I'm hoping to restructure how tests are added.  As of this
writing, 5 of the tests fail (intentionally), but I think I will
make it so that they "pass" as far as pytest is concerned.