Checking in jabmod transformer (which may replace much of the code here).
It would seem that Claude 3.7 did a better job figuring out Lark than
I have over the past few years. Most of ABIFtoJabmodTransformer was
generated by Claude with my direction and tweaks along the way.  The
jabmod looks like it should be compatible with the abiftool, but I
haven't tested that (yet).